fastapi_django_main_lives/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
gradio==4.31.5
starlette==0.36.3
pydantic==2.6.2
pydantic-core==2.16.3

# Django
django==5.0.4
asgiref==3.8.1
sqlparse==0.5.0
whitenoise==6.6.0

# AI & ML
open-interpreter
groq==0.4.1
llamafactory
accelerate
diffusers

# Database
duckdb
psycopg2-binary

# Utilities
python-dotenv
jinja2
aiofiles
click==8.1.7
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
sniffio==1.3.1
sse-starlette==2.0.0
typing-extensions==4.10.0
colorama==0.4.6
exceptiongroup==1.2.0
tzdata==2024.1
uvloop==0.19.0
huggingface-hub
imageio[ffmpeg]
torch
torchvision
transformers>=4.41.2
langchain
langchain_groq
sqlalchemy
sentence-transformers
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
line-bot-sdk
gradio_client
